Added Relations

+++ tax-app-backend/database/migrations/2014_10_12_000000_create_users_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
-            $table->integer('role_id');
+            $table->unsignedBigInteger('role_id');
             $table->string('latitude');
             $table->string('longitude');
             $table->string('image_url');
